[ADD] remove old pdf gen for the final on
